requirements.txt
Django==5.1.4
django-cors-headers==4.6.0
mysqlclient==2.2.6
python-decouple==3.8
